backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-jose
passlib[bcrypt]
pydantic
alembic
python-dotenv
requests